requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
google-api-python-client==2.118.0
requests==2.31.0
pydantic==2.6.1
python-multipart==0.0.9
google-generativeai==0.7.1
selenium==4.18.1
undetected-chromedriver==3.5.5
2captcha-python==1.2.0
tzdata>=2023.3